Rename and update rectangle_around usage examples
Renamed usage example files from 'rectangle_around-1-example' to 'rectangle_around_circle-1-example' across multiple languages for clarity. Updated usage-example references and added a new entry for 'rectangle_around_circle' in usage-example-references.json. Improved fallback logic in getGroupName in api-pages-script.cjs to format group names from uniqueName if no match is found.
