+++ public/usage-examples/geometry/rectangle_around_circle-1-example.gif
